requirements.txt
flask==3.1.0
requests==2.32.3
beautifulsoup4==4.13.4
pandas==2.2.3
#nltk==3.8.1
#tiktoken==0.5.1
python-dotenv==1.1.0
openai==2.14.0
selenium==4.31.0 
validators==0.35.0
uvicorn==0.40.0
fastapi==0.128.0
boto3==1.38.0
pillow==12.0.0
python-multipart==0.0.21
Flask-SQLAlchemy==3.1.1
boto3-stubs[dynamodb,sqs]==1.42.19
